move demo survey to components
