+++ ui/src/app/pages/login/account-creation/account-creation-survey.tsx
@@ -4,7 +4,7 @@ import * as fp from 'lodash/fp';
 import { Profile } from 'generated/fetch';
 
 import { environment } from 'environments/environment';
-import { DemographicSurvey } from 'app/pages/profile/demographic-survey';
+import { DemographicSurvey } from 'app/components/demographic-survey';
 import { profileApi } from 'app/services/swagger-fetch-clients';
 import { AnalyticsTracker } from 'app/utils/analytics';
