+++ ui/src/app/pages/profile/demographic-survey-panel.tsx
